implement formatting for arrows and test effect formatting
