+++ main/src/ca/uwaterloo/flix/language/debug/Audience.scala
@@ -1,0 +1,7 @@
+package ca.uwaterloo.flix.language.debug
+
+sealed trait Audience
+object Audience {
+  case object Internal extends Audience
+  case object External extends Audience
+}
